pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Sustavov Maksim <maksim.sustavov@edgecenter.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.0"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
validators = "^0.34.0"
psycopg2-binary = "^2.9.10"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
isort = "^5.13.2"
flake8-isort = "^6.1.1"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
